+++ src/main/java/rentalstore/statement/RentalStatement.java
@@ -13,7 +13,7 @@ public class RentalStatement extends Statement {
     }
 
     @Override
-    protected String getFooterString(double totalAmount, int frequentRenterPoints) {
+    protected String getFooterString(double totalAmount, double frequentRenterPoints) {
         String result = "Amount owed is " + String.valueOf(totalAmount) + "\n";
         result += "You earned " + String.valueOf(frequentRenterPoints) + " frequent renter points";
         return result;
